fix(tt_schedule): widen snapshot offset to uint32_t + strengthen stale test
tt_freeze_inputs/bm_let_in 用 uint8_t 累加输入快照偏移，合法配置下
(MAX_INPUTS=8 x MAX_ELEM_SIZE=64) 累计偏移可超 255 静默回绕，导致不同
输入快照互相踩踏。改为 uint32_t 并去掉截断转型。

test_freeze_stale_when_never_published 补两条断言：age 哨兵值
0xFFFFFFFFu、冻结快照被填成对应 safe_default，覆盖 WOULD_BLOCK 兜底路径
的完整契约。

+++ Source/hybrid/bm_tt_schedule.c
@@ -49,7 +49,7 @@ struct bm_let_ctx {
  */
 static void tt_freeze_inputs(bm_tt_schedule_t *s, bm_tt_activity_t *a) {
     uint32_t period_us = s->minor_us * a->every;
-    uint8_t off = 0u;
+    uint32_t off = 0u;
 
     for (uint8_t i = 0u; i < a->input_count; ++i) {
         const bm_let_input_t *in = &a->inputs[i];
@@ -75,7 +75,7 @@ static void tt_freeze_inputs(bm_tt_schedule_t *s, bm_tt_activity_t *a) {
             a->rt->age_us[i] = age;
             a->rt->stale[i] = (max_age != 0u && age > max_age) ? 1 : 0;
         }
-        off += (uint8_t)in->elem_size;
+        off += in->elem_size;
     }
 }
 
@@ -85,10 +85,10 @@ static void tt_freeze_inputs(bm_tt_schedule_t *s, bm_tt_activity_t *a) {
 const void *bm_let_in(bm_let_ctx_t *ctx, uint32_t in_idx, int *out_stale,
                        uint32_t *out_age_us) {
     const bm_tt_activity_t *a = ctx->act;
-    uint8_t off = 0u;
+    uint32_t off = 0u;
 
     for (uint32_t i = 0u; i < in_idx; ++i) {
-        off += (uint8_t)a->inputs[i].elem_size;
+        off += a->inputs[i].elem_size;
     }
     if (out_stale) {
         *out_stale = a->rt->stale[in_idx];
